+++ scripts/mcp-tool-pack-baseline.json
@@ -10,11 +10,9 @@
   "approve_decomposition",
   "assess_archetype_refinement",
   "assess_complexity",
-  "assess_contribution",
   "check_deployment_windows",
   "check_sandbox",
   "confirm_taxonomy_placement",
-  "contribute_to_hive",
   "create_backlog_item",
   "create_build_epic",
   "create_digital_product",
@@ -55,8 +53,6 @@
   "propose_build_decomposition",
   "propose_decomposition",
   "propose_file_change",
-  "propose_improvement",
-  "propose_skill_improvement",
   "provision_build_engine",
   "query_backlog",
   "read_sandbox_file",
@@ -94,7 +90,6 @@
   "start_ideate_research",
   "start_sandbox",
   "start_scout_research",
-  "submit_feedback",
   "suggest_taxonomy_placement",
   "summon_coworker",
   "triage_backlog_item",
